requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.36.0
anyio==4.5.2
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
crispy-bootstrap5==2024.10
cryptography==43.0.1
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
Django==5.1.1
django-allauth==65.0.2
django-axes==7.0.0
django-background-tasks==1.2.8
django-ckeditor==6.7.1
django-ckeditor-5==0.2.14
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-formtools==2.5.1
django-import-export==4.1.1
django-ipware==7.0.1
django-js-asset==2.2.0
django-microsoft-auth==3.0.1
django-model-utils==5.0.0
django-redis==5.4.0
django-sslserver==0.22
django-storages==1.14.4
django-taggit==6.1.0
executing==2.1.0
Faker==33.3.0
fastjsonschema==2.20.0
filelock==3.16.1
flake8==6.1.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
geoip2==4.8.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
granian==1.6.3
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
maxminddb==2.6.2
mccabe==0.7.0
mistune==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.51.2
openai-whisper==20240930
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.1.0
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
python-json-logger==2.0.7
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tablib==3.5.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.1
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
uvloop==0.21.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.7.0
yarl==1.13.1
yt-dlp==2024.11.4
